- Fixed: The C code in AltSoundRenderer::CopyAndClip() did not shift the sample   data enough (2 bits instead of 8), so it was super loud and aliased. - Fixes for GCC 4.1: Several type-punned pointer warnings, but more   importantly, declaring a friend function inside a class body is no longer   enough to declare that function globally; you must declare it again outside   the class. - Upgraded FArchive::SerializePointer so that it can store 32-bit indices. - ACS printing pcodes now build their string in an FSttring instead of a fixed   sized buffer on the stack.
SVN r145 (trunk)

+++ src/win32/i_input.cpp
@@ -1745,20 +1745,20 @@ static void MouseRead_DI ()
 		/* Look at the element to see what happened */
 		// GCC does not like putting the DIMOFS_ macros in case statements,
 		// so use ifs instead.
-		if (od.dwOfs == DIMOFS_X)
+		if (od.dwOfs == (DWORD)DIMOFS_X)
 		{
 			dx += od.dwData;
 		}
-		else if (od.dwOfs == DIMOFS_Y)
+		else if (od.dwOfs == (DWORD)DIMOFS_Y)
 		{
 			dy += od.dwData;
 		}
-		else if (od.dwOfs == DIMOFS_Z)
+		else if (od.dwOfs == (DWORD)DIMOFS_Z)
 		{
 			WheelMove += od.dwData;
 			WheelMoved ();
 		}
-		else if (od.dwOfs >= DIMOFS_BUTTON0 && od.dwOfs <= DIMOFS_BUTTON7)
+		else if (od.dwOfs >= (DWORD)DIMOFS_BUTTON0 && od.dwOfs <= (DWORD)DIMOFS_BUTTON7)
 		{
 			/* [RH] Mouse button events mimic keydown/up events */
 			if (!GUICapture)
